refactor: rename common directory to abstractions
