+++ Src/Core/Domain/Entities/PlaceAddress.cs
@@ -1,4 +1,4 @@
-﻿using Domain.Entities.Common;
+﻿using Domain.Entities.Abstractions;
 
 namespace Domain.Entities;
