+++ Src/Core/Domain/Interfaces/Persistence/Repositories/Abstractions/IRepository.cs
@@ -1,0 +1,6 @@
+﻿namespace Domain.Interfaces.Persistence.Repositories.Abstractions;
+
+public interface IRepository
+{
+
+}
