--- Src/Core/Domain/Interfaces/Persistence/Repositories/Common/IRepository.cs
@@ -1,6 +1,0 @@
-﻿namespace Domain.Interfaces.Persistence.Repositories.Common;
-
-public interface IRepository
-{
-
-}
