+++ Src/Core/Domain/Interfaces/Persistence/Repositories/Read/Abstractions/IReadPersonRoleEntityRepository.cs
@@ -1,6 +1,6 @@
-﻿using Domain.Entities.Common;
+﻿using Domain.Entities.Abstractions;
 
-namespace Domain.Interfaces.Persistence.Repositories.Read.Common;
+namespace Domain.Interfaces.Persistence.Repositories.Read.Abstractions;
 
 public interface IReadPersonRoleEntityRepository<TPerson> : IReadRepositoryBase<TPerson>
     where TPerson : PersonRoleEntityBase
