--- Src/Core/Domain/Interfaces/Persistence/Repositories/Read/Common/IReadRepositoryBase.cs
@@ -1,10 +1,0 @@
-﻿using Domain.Entities.Common;
-using Domain.Interfaces.Persistence.Repositories.Common;
-
-namespace Domain.Interfaces.Persistence.Repositories.Read.Common;
-
-public interface IReadRepositoryBase<TEntity> : IRepository
-    where TEntity : EntityBase
-{
-    public TEntity? Get(long id);
-}
