+++ Src/Core/Domain/Interfaces/Persistence/Repositories/Write/Common/IWriteEntityRepositoryBase.cs
@@ -1,7 +1,7 @@
-﻿using Domain.Entities.Common;
-using Domain.Interfaces.Persistence.Repositories.Common;
+﻿using Domain.Entities.Abstractions;
+using Domain.Interfaces.Persistence.Repositories.Abstractions;
 
-namespace Domain.Interfaces.Persistence.Repositories.Write.Common;
+namespace Domain.Interfaces.Persistence.Repositories.Write.Abstractions;
 
 public interface IWriteEntityRepositoryBase<TEntity> : IRepository
     where TEntity : EntityBase
