+++ Src/Core/Domain/Interfaces/Persistence/Repositories/Write/IWriteCustomersRepository.cs
@@ -1,5 +1,5 @@
 ﻿using Domain.Entities;
-using Domain.Interfaces.Persistence.Repositories.Write.Common;
+using Domain.Interfaces.Persistence.Repositories.Write.Abstractions;
 
 namespace Domain.Interfaces.Persistence.Repositories.Write;
