+++ Src/Infrastructure/Persistence/Repositories/Abstractions/RepositoryBase.cs
@@ -1,8 +1,8 @@
-﻿using Domain.Entities.Common;
+﻿using Domain.Entities.Abstractions;
 using Microsoft.EntityFrameworkCore;
 using Persistence.DbContexts;
 
-namespace Persistence.Repositories.Common;
+namespace Persistence.Repositories.Abstractions;
 
 internal abstract class RepositoryBase<TEntity>
 where TEntity : EntityBase
@@ -10,7 +10,7 @@ where TEntity : EntityBase
     protected readonly ReadDbContext _context;
     protected readonly DbSet<TEntity> _set;
 
-    public RepositoryBase(ReadDbContext context)
+    protected RepositoryBase(ReadDbContext context)
     {
         _context = context;
         _set = context.Set<TEntity>();
