+++ Src/Infrastructure/Persistence/Repositories/Read/Abstractions/ReadPersonRoleRepository.cs
@@ -1,9 +1,9 @@
-﻿using Domain.Entities.Common;
-using Domain.Interfaces.Persistence.Repositories.Read.Common;
+﻿using Domain.Entities.Abstractions;
+using Domain.Interfaces.Persistence.Repositories.Read.Abstractions;
 using Domain.ValueObjects;
 using Persistence.DbContexts;
 
-namespace Persistence.Repositories.Read.Common;
+namespace Persistence.Repositories.Read.Abstractions;
 
 internal abstract class ReadPersonRoleRepository<TPerson> : ReadRepositoryBase<TPerson>, IReadPersonRoleEntityRepository<TPerson>
     where TPerson : PersonRoleEntityBase
