+++ Src/Infrastructure/Persistence/Repositories/Read/ReadCustomersRepository.cs
@@ -1,7 +1,7 @@
 ﻿using Domain.Entities;
 using Domain.Interfaces.Persistence.Repositories.Read;
 using Persistence.DbContexts;
-using Persistence.Repositories.Read.Common;
+using Persistence.Repositories.Read.Abstractions;
 
 namespace Persistence.Repositories.Read;
