+++ Src/Infrastructure/Persistence/Repositories/Write/WriteCustomerRepository.cs
@@ -1,7 +1,7 @@
 ﻿using Domain.Entities;
 using Domain.Interfaces.Persistence.Repositories.Write;
 using Persistence.DbContexts;
-using Persistence.Repositories.Write.Common;
+using Persistence.Repositories.Write.Abstractions;
 
 namespace Persistence.Repositories.Write;
