+++ Tests/Infrastructure/PersistenceTests/GlobalUsings.cs
@@ -1,1 +1,2 @@
 ﻿global using Moq;
+global using FluentAssertions;
