+++ Tests/Infrastructure/PersistenceTests/Repositories/Read/Abstractions/IReadPersonRoleRepositoryTests.cs
@@ -1,4 +1,4 @@
-﻿namespace PersistenceTests.Repositories.Read.Common;
+﻿namespace PersistenceTests.Repositories.Read.Abstractions;
 
 public interface IReadPersonRoleRepositoryTests : IReadRepositoryBaseTests
 {
